fix: redux was being started in dev mode in production
We weren’t properly passing through the environment, and it defaulted to development.  Now it defaults to production.

Also fixing two linting errors.
